requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-mistralai==1.1.1
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.1
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pgvector==0.4.2
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdf==6.6.2
python-dotenv==1.2.1
lingua-language-detector==2.0.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
SQLAlchemy==2.0.46
starlette==0.50.0
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
